move thing around
