+++ data-processing-service/src/main/java/gov/cdc/dataprocessing/service/interfaces/IPatientService.java
@@ -4,9 +4,10 @@ import gov.cdc.dataprocessing.exception.DataProcessingConsumerException;
 import gov.cdc.dataprocessing.exception.DataProcessingException;
 import gov.cdc.dataprocessing.model.classic_model.dt.EdxLabInformationDT;
 import gov.cdc.dataprocessing.model.classic_model.vo.LabResultProxyVO;
+import gov.cdc.dataprocessing.model.classic_model.vo.PersonVO;
 
 public interface IPatientService {
-    Object processingPatient(LabResultProxyVO labResultProxyVO, EdxLabInformationDT edxLabInformationDT) throws DataProcessingConsumerException, DataProcessingException;
+    Object processingPatient(LabResultProxyVO labResultProxyVO, EdxLabInformationDT edxLabInformationDT, PersonVO personVO) throws DataProcessingConsumerException, DataProcessingException;
     Object processingNextOfKin() throws DataProcessingConsumerException;
     Object processingProvider() throws DataProcessingConsumerException;
 }
