+++ data-processing-service/src/main/java/gov/cdc/dataprocessing/utilities/component/patient/EdxPatientMatchRepositoryUtil.java
@@ -1,4 +1,4 @@
-package gov.cdc.dataprocessing.utilities.component;
+package gov.cdc.dataprocessing.utilities.component.patient;
 
 import gov.cdc.dataprocessing.model.classic_model.dto.EdxPatientMatchDT;
 import gov.cdc.dataprocessing.repository.nbs.odse.EdxPatientMatchRepository;
