+++ data-processing-service/src/main/java/gov/cdc/dataprocessing/utilities/component/patient/EdxPatientMatchingCriteriaUtil.java
@@ -1,4 +1,4 @@
-package gov.cdc.dataprocessing.utilities.component;
+package gov.cdc.dataprocessing.utilities.component.patient;
 
 import gov.cdc.dataprocessing.constant.elr.EdxELRConstant;
 import gov.cdc.dataprocessing.constant.elr.NEDSSConstant;
