+++ data-processing-service/src/main/java/gov/cdc/dataprocessing/utilities/component/patient/EdxPatientMatchingHelper.java
@@ -1,4 +1,4 @@
-package gov.cdc.dataprocessing.utilities.component;
+package gov.cdc.dataprocessing.utilities.component.patient;
 
 import gov.cdc.dataprocessing.exception.DataProcessingException;
 import gov.cdc.dataprocessing.model.classic_model.dto.EdxPatientMatchDT;
